backend/requirements.txt
google-generativeai
pandas
openpyxl
python-dotenv
fastapi
uvicorn[standard]
pydantic-settings
psycopg2-binary
composio
